Filterpage update and functionality from mainscreen
